+++ darwin/Classes/Logger.swift
@@ -1,0 +1,7 @@
+import Foundation
+import os.log
+
+extension Logger {
+    private static var subsystem = Bundle.main.bundleIdentifier!
+    public static let main = Logger(subsystem: subsystem, category: "main")
+}